Changed the plugin filename to libfribidplugin.so
This makes it possible to install Nexus Personal and FriBID
side-by-side. The only problem is that the plugin filename is exposed to
the web sites, which could "sniff" for certain filenames. However, the
plugin description is also exposed (and it's of course different from
the one in Nexus Personal). Furthermore, I haven't seen any sites that
sniffs for anything else than the MIME type, so I think this change is
motivated.