Provide a secure memory pool and use it
This implements a secure memory pool for use in fribid, allocating
a page at a time for secure use. We currently only use one page
for the passphrase, but an arbitrary number of pages can be made
available. We currently don't need more intelligence than this.

The page pool is mmap():ed in to make sure it's on even pages,
and the entire pool is then mlock():ed to hinder it from being
spooled out as swap.

This also augments the platform_sign function so that it takes the
piece of (secure) memory used to store the password as a parameter,
and copied the passphrase into it as soon as it's retrieved, so
that it is not allocated or passed around elsewhere.

Signed-off-by: Linus Walleij <linus.ml.walleij@gmail.com>