Fix some bugs from the second last patch
These changes fix two bugs from the secure memory patch,
2605e8dd6ffd0b8cc710a3f91d58192cd510473a

The one in gtk.c was caused by the password variable being
dereferenced after it had been indexed (an operator precedence problem).
This could cause a crash.

The one in main.c was a logic error were an error message was sent
as a parameter to the plugin. However, the plugin expects this parameter
to be the Signature so this was wrong (but hard to spot!).

I also changed a secmem_init_pool function to reset the pool pointer to
NULL if mmap fails, to avoid errors later on.