Convert the code to use OpenSSL instead of NSS
This commit is an exact copy of Marcus Carlsons "work in progress" patch.

The main reason for using OpenSSL instead of NSS is that the design of
NSS is too centered around it's database and secure cryptographic
domains. This doesn't work well with the BankID protocol because it
stores the certificates as files. -- Samuel