Add an abstraction layer for tokens
This patch adds a common interface for different kinds of tokens. The
interface should hopefully be able to support both smart cards and
file-based tokens (PKCS12 and NGE/NGP), but at the moment only PKCS12
tokens are implemented.

I've included a few other changes in this patch as well because it was
easier to make them at the same time. This includes more useful error
messages and a little refactoring of the PKCS12 code.