Increase the minimum GTK version to 2.14 to avoid problems
If FriBID is built with GTK 2.14 or later it can avoid direct access to
GTK's data structures, and use accessor functions instead. This will
be a requirement for all GTK 3.0 applications. However, this also means
that if FriBID is built with 2.14 or later it won't work with older
versions of GTK.

This patch increases the minimum GTK version in the Debian build files
to avoid this problem.