Use shlibs:Depends in the Debian package
Previously FriBID supported some old platforms like Ubuntu 8.04, so the
"Depends:" line had to be specified manually because shlibs was too
picky with the dependencies (even when the API documentations explicitly
said that older versions should work and when they also did work).

However, this caused problems when FriBID was built on some old
platforms like Debian Lenny with Smartcard support (the default). This
patch changes the Debian control file to use ${shlibs:Depends} which
solves the latter problem.