Check the correct parameters in SetParam
This error was introduced by commit
f2f771cf891abac6120060bd52c3b879f0620f92 . It hasn't been included in
any stable release.

SetParam calls IS_CALL_2, which in turn call ARG which has a bug. This
could cause problems if the second parameter is not a string and the
code casts it to a string pointer and a length value. It does survive
the input fuzzer, but it might crash or leak data with carefully crafted
data.