Abort if no requests are passed to createRequest
This works fine without this patch (= returns an error code), but this
happens just because the PKCS12 code returns an error when there aren't
any keys. I plan to make some changes that use the name attribute from
the first certificate before that point, so this needs to be changed.