Check if decoded strings are UTF-8 or ISO-8859-1 or invalid
With this commit ISO-8859-1 and invalid strings are handled properly.
ISO-8859-1 strings are converted to UTF-8 and invalid strings are
rejected and cause an error code to be returned.

Strings with invalid characters isn't just a cosmetic problem. GTK
rejects strings that are not valid UTF-8 completely (but control
characters are allowed and are displayed in hexadecimal, so I though it
handled *any* binary data when I first wrote the code). If the invalid
UTF-8 strings are sent to the plugin then "text to sign" will be
empty, but the text will still be included in the signature (with
invalid UTF-8).