Fix warning messages from "make clean" when config.h doesn't exist
The Makefiles read some information from common/config.h, but when this
file doesn't exist (for example if make clean is run before ./configure)
then this will cause an error message to be displayed.

This patch hides the error message because the Makefile rules that
need config.h, depend on a rule that prints an error message if this
file is not found.