Remove version expiry status checking system
This system had two purposes:
  1) to make FriBID able to determine the latest version of the official
     software it could emulate, without having to upgrade FriBID itself.
  2) to make it possible to "revoke" old versions of FriBID in case of
     security problems.

It has turned out that this system is not necessary. 1) was never
really needed because the official software was updated quite rarely
anyway and only a few web sites cared about the version string anyway.
2) is a kind of "kill switch" but it's not very useful since it's
only checked when you use the FriBID user interface. Hence users who
use FriBID (the user interface) infrequently or users who do not use
it any longer would not see the warning message. And it seems that most
users only use FriBID infrequently (typically once a year or so).

This change removes a lot of code. One of the function calls removed
is platform_seedRandom which was ONLY used for the "expiry" feature
which used rand(). This does not affect the crypto stuff which uses
OpenSSL's random generator which in turn uses /dev/random.