Build with -fPIE/-pie and -lpthread
The former is needed for further security hardening which may be done by
the kernel (e.g. ASLR) or the distribution (e.g. Debian adding extra
hardenings only when PIE is enabled). The latter is needed in case the
pthread functions are not available from the other libraries or a dependency
of them.